updated nonces to bigInt

+++ lib/iden3comm/data/repositories/iden3comm_repository_impl.dart
@@ -166,7 +166,7 @@ class Iden3commRepositoryImpl extends Iden3commRepository {
   @override
   Future<Uint8List> getAuthInputs(
       {required String genesisDid,
-      required int profileNonce,
+      required BigInt profileNonce,
       required String challenge,
       required List<String> authClaim,
       required IdentityEntity identity,
